fix: 닉네임 변경 시 문자열 길이 최대값 제한 추가, chore: auth 관련 utils 함수 정리 중
